[project @ Update user colour in key press event handler if possible [fixes #24]] If someone copied a part of the buffer, the tags applied to it were copied, too. If he pasted it then, the insert-text signal handler could not remove those tags because they were not yet applied. So it ended up in having to user colour tags applied to the new range of text: The one from the user who pasted the text and the one from the user who wrote the original text.
This is solved by adding tags followed by a key press event not anymore in the
insert-text event handler but in the key press event handler, which is called
later, when the pasted tag already has been applied. Then, it may be removed and
the one and only user colour tag can be set.

Original author: Armin Burgmeier <armin@0x539.de>
Date: 2005-06-04 09:52:11+00:00