Fix a possible crash when adding a user to the userlist
A crash could happen if the cell data func was called before the user
object was set in the row of the list store.

A cleaner way to avoid this would be to atomically create the row and set
the data, using gtk_list_store_insert_with_values(), but this API is not
wrapped in gtkmm.

Many thanks to Rainer Rehak for reporting this!
(http://permalink.gmane.org/gmane.network.obby.user/411)