Use OperationSubscribePath for direct connection handling
This simplifies the code in the Browser class a bit, and does exactly the
same thing. Also, it enables a new feature which is to write the full path
to a document on the remote server into the direct connection field, and
Gobby then tries to subscribe to the document.