Fix a crash switching between disconnected documents
This resulted from the switch to the InfBrowser interface in libinfinity. The
InfBrowser::unsubscribe-signal is now also emitted on disconnection or when
the corresponding browser node is removed and the session was subscribed.
However, this does not mean that we were unsubscribed from the session itself.
For the latter, we watch InfSession::notify::subscription-group, which we now
use as the only means for detecting unsubscription.

Previously, it could happen that InfBrowser::unsubscribe-session was emitted
in which case we removed the signal handler for
InfSession::notify::subscription-group, but did not reset the active user from
the session view, and then switching to the document lead to a crash.