Refactor the folder management
Introduce a new class FolderManager which takes care of TextFolder
and ChatFolder. The FolderManager can be used to add new documents into
the appropriate folder, and it provides add/remove signals that give
access to the session proxy.

This allows UserJoinCommands to use the proxy only relying on core objects,
and it allows creating documents only for those subscriptions that were
requested by the local client.