Fix a crash when subscribing to a local document that has remote subscriptions
This crash is strange since it happens in
gtk_source_buffer_set_style_scheme() when the TextSessionView object is
created. There is one subtle difference to the other two cases:

 * Subscribing to a remote document
 * Subscribing to a local document that has no remote subscriptions

which is that the GtkSourceBuffer is newly created; while if there are
remote subscriptions, then the buffer exists already and the style
scheme is set on a buffer that is already in use for a while.

Now, I don't see why this would produce a crash. I have moved the
setting of the style scheme further up, before the buffer is connected
to the GtkSourceView, and the problem seems to be gone. However, I
suspect a deeper problem down in GTK+ somewhere...