Prevent session closure of local documents when opening preferences dialog
The reason for the problem was that the "selection-changed" signal of
GtkFileChooserButton is first emitted with the home directory and then
again with the actual initial value. This caused the storage of the
InfdDirectory to be reset temporarily.

In principle, this can be fixed by using the "file-set" signal instead.
However, a signal emission is missing when selecting an entry from the
combo box instead of the file dialog. This was only fixed in GTK+ 3.10
or later. Therefore, a workaround is implemented which uses the
"selection-changed" signal but ignores it until it was first emitted
with the initial value.