Remove the option not to generate a certificate in the initial dialog
This was broken already before, since it would need the security policy
to be set to only-unsecured, which is not possible from within the Gobby UI.
We don't want to make it available, because otherwise it would also disable
TLS for outgoing connections. Having two different settings for incoming
and outgoing connections seems messy, so we simply always generate a
certificate.