metainfo: Add OARS age-rating data
This data is used by software centre applications such as GNOME Software.
The specification and generator are at https://hughsie.github.io/oars/.

Walking through the generator, the only relevant attribute is
'social-chat':

> Online Text-only Messaging
> --------------------------
>
> Defined as any messaging system connected to the Internet.

The 'intense' value (the highest level) is defined as:

> Uncontrolled chat functionality between users.

Gobby is a network-connected text editor so its entire purpose is
indeed uncontrolled text communication between users over a network or
the internet.

At the time of writing, this rating causes libappstream to assign the
app a minimum age of 13 years.